Add support for a GenericWorkflowWorker
GenericWorkflowWorker allows specifying a function that can use any
method it wants to lookup workflow definitions from a name and version,
including by dynamically  it.

To support this, the threading and multiprocessing based workers are
renamed "Executor" and accept a worker as an argument so that they can
be used with either the normal WorkflowWorker or GenericWorkflowWorker.